requirements.txt
alembic==1.8.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.30.1
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
bcrypt==4.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
click-params==0.3.0
cloudpickle==2.2.1
cmdstanpy==1.2.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docker==6.1.3
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.1
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
holidays==0.47
httpcore==1.0.5
httplib2==0.19.1
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
ipython==8.23.0
ipywidgets==8.1.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.4.0
numpy==1.26.4
orjson==3.10.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
prophet==1.1.5
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pymongo==4.6.3
PyMySQL==1.0.3
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
stack-data==0.6.3
stanio==0.5.0
starlette==0.37.2
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
validators==0.18.2
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.10
zenml==0.56.3
